+++ test-parts/part-i-api_frontend/parts/a-check_database_empty.js
@@ -9,7 +9,7 @@ const apiRequestScript = require(commonPaths.requestApi);
 const httpRequests = require(commonPaths.httpRequests);
 const rioCommon = require(commonPaths.rioCommon);
 const modelFunctionsFile = require(commonPaths.getModels);
-const modelArray = modelFunctionsFile.retrieveAllSupportedModels();
+const modelArray = modelFunctionsFile.getAllModels();
 
 var retrievedDatabaseList = null;
 
@@ -59,7 +59,7 @@ function handleDatabaseCheck()
 		it("Valid Array Structure", function(done)
 		{
 			commonFunctionsFile.testAllElements(retrievedDatabaseList, 'object');
-			rioCommon.callTestDeviceArrayStructure(retrievedDatabaseList);
+			rioCommon.testDeviceArray(retrievedDatabaseList);
 			done();
 		});